LTI-121 Renamed tool from roster-management to section-management, changed file names to reflect the new name
