+++ .gitignore
@@ -16,4 +16,6 @@ buildNumber.properties
 # JDT-specific (Eclipse Java Development Tools)
 .classpath
 
+.idea
+
 *.tar
