+++ README.md
@@ -1,1 +1,1 @@
-# roster-management
+# section-management
