--- src/main/java/edu/virginia/its/canvas/roster/repos/WaitlistedSectionRepo.java
@@ -1,9 +1,0 @@
-package edu.virginia.its.canvas.roster.repos;
-
-import edu.virginia.its.canvas.roster.model.WaitlistedSection;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WaitlistedSectionRepo extends JpaRepository<WaitlistedSection, Long> {
-  List<WaitlistedSection> findBySisSectionIdIn(List<String> sections);
-}
