+++ src/main/java/edu/virginia/its/canvas/section/AuditConfig.java
@@ -1,0 +1,20 @@
+package edu.virginia.its.canvas.section;
+
+import java.time.OffsetDateTime;
+import java.util.Optional;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.auditing.DateTimeProvider;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@Configuration
+@EnableTransactionManagement
+@EnableJpaAuditing(dateTimeProviderRef = "dateTimeProvider")
+public class AuditConfig {
+
+  @Bean(name = "dateTimeProvider")
+  public DateTimeProvider dateTimeProvider() {
+    return () -> Optional.of(OffsetDateTime.now());
+  }
+}
