+++ src/main/java/edu/virginia/its/canvas/section/model/SectionManagementForm.java
@@ -1,0 +1,17 @@
+package edu.virginia.its.canvas.section.model;
+
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public final class SectionManagementForm {
+
+  private String selectedTerm;
+  private List<String> sectionsToKeep = new ArrayList<>();
+  private List<String> sectionsToAdd = new ArrayList<>();
+  private List<String> waitlistsToAdd = new ArrayList<>();
+  private boolean removeFromCourse;
+}
