+++ src/main/java/edu/virginia/its/canvas/section/utils/Constants.java
@@ -1,0 +1,9 @@
+package edu.virginia.its.canvas.section.utils;
+
+public final class Constants {
+  private Constants() {}
+
+  public static final String CANVAS_LINK_HEADER = "link";
+  public static final String USERNAME_CUSTOM_KEY = "username";
+  public static final String COURSE_ID_CUSTOM_KEY = "course_id";
+}
