+++ src/main/resources/static/js/section.js
