+++ src/main/resources/templates/fragments/head.html
@@ -8,7 +8,7 @@
     <title>Manage SIS Sections</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous" />
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.4/dist/jquery.min.js"></script>
-    <script type="text/javascript" th:src="@{/js/roster.js}"></script>
-    <link th:href="@{/css/roster.css}" rel="stylesheet"/>
+    <script type="text/javascript" th:src="@{/js/section.js}"></script>
+    <link th:href="@{/css/section.css}" rel="stylesheet"/>
   </head>
 </html>
