+++ src/main/resources/templates/waitlists.html
@@ -8,7 +8,7 @@
 
 <body>
 <div th:replace="fragments/header :: header"></div>
-<form action="#" th:action="@{/validate}" th:object="${rosterManagementForm}" method="post">
+<form action="#" th:action="@{/validate}" th:object="${sectionManagementForm}" method="post">
   <input type="hidden" th:field="*{sectionsToAdd}"/>
   <input type="hidden" th:field="*{sectionsToKeep}"/>
   <h4>Manage Waitlisted Students</h4>
