--- src/test/java/edu/virginia/its/canvas/roster/controller/RosterManagementControllerTest.java
@@ -1,10 +1,0 @@
-package edu.virginia.its.canvas.roster.controller;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class RosterManagementControllerTest {
-  @Test
-  void contextLoads() {}
-}
