+++ src/test/java/edu/virginia/its/canvas/section/controller/SectionManagementControllerTest.java
@@ -1,0 +1,10 @@
+package edu.virginia.its.canvas.section.controller;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class SectionManagementControllerTest {
+  @Test
+  void contextLoads() {}
+}
